Complete details page with bottom sheet to display character comics/series/stories
- Create generic marvel response model
- Use router to handle navigation

+++ Marvelpedia/Modules/Characters/Common/Business/Api/Endpoints/CharacterEndpoints.swift
@@ -9,7 +9,7 @@ import Foundation
 
 enum CharacterEndpoint: NetworkEndpoint {
     case list
-    case details(characterId: Int)
+    case details(characterId: Double)
     
     func getEndpoint() -> String {
         switch self {
